requirements.txt
apache-airflow
pandas
openpyxl
psycopg2-binary
sqlalchemy
pytest
pyyaml